Fixed edit task

+++ backend/routes/tasksRouter.js
@@ -97,10 +97,15 @@ router.put("/", isAuthenticated, async (req, res) => {
         if(!task.exists){
             throw Error("Task not found")
         }
+        if(task.user !== req.userRef){
+            throw Error("Task unaccessible")
+        }
         const deletedAttachments = task.attachments.filter((attachment) => !attachments.includes(attachment))
         await Promise.all(deletedAttachments.map(async (attachment) => { return await deleteMediaObject(attachment) }))
 
         await updateDocument("tasks", {
+            user: task.user,
+            createdAt: task.user.createdAt,
             ...(title ? { title, titleFuzzy: generateFuzzyArray(title) } : {}),
             ...(description ? { description, descriptionFuzzy: generateFuzzyArray(description) } : {}),
             ...(status ? { status } : {}),
@@ -128,6 +133,9 @@ router.delete("/", isAuthenticated, async (req, res) => {
         const task = await getDocumentWithId("tasks", id);
         if(!task.exists){
             throw Error("Task not found")
+        }
+        if(task.user !== req.userRef){
+            throw Error("Task unaccessible")
         }
         await Promise.all(task.attachments.map(async (attachment) => { return await deleteMediaObject(attachment) }))
